backend/requirements.txt
audioop-lts==0.2.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.12
Flask-Session>=0.5.0
fsspec==2025.3.0
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
networkx==3.4.2
numpy==2.2.4
ordered-set==4.1.0
packaging==24.2
pydub==0.25.1
Pygments==2.19.1
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
setuptools==77.0.3
SpeechRecognition==3.14.1
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug>=2.3.0
wrapt==1.17.2
psutil==5.9.0
Flask-Migrate>=4.0.5
Flask-SQLAlchemy>=3.1.1
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
requests>=2.31.0
oauthlib>=3.2.2
pyOpenSSL>=23.3.0
cryptography>=41.0.0
